fix: news page and code styling
