+++ app/mocks/fileMock.js
@@ -1,1 +1,1 @@
-module.exports = {}
+module.exports = {};
