+++ app/src/Components/Home/CardsList.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import Card from './Card';
 
 const CardsList = (props) => {
-
   const { cards } = props;
 
   const CARDS_LIST = [];
