+++ app/tests/react.test.js
@@ -1,22 +1,19 @@
 import React from 'react';
+import { describe, it } from '@jest/globals';
+import { render } from '@testing-library/react';
 import App from '../src/App';
 import About from '../src/Components/About/About';
-import {describe, it} from "@jest/globals";
-import { render } from '@testing-library/react';
 
-
-describe ('testing react components', ()=>{
-
-  it('app', ()=>{
-    let app = render(App());
-    let el = app.getByTestId("App");
-    expect(el).toBeInstanceOf(HTMLElement)
+describe('testing react components', () => {
+  it('app', () => {
+    const app = render(App());
+    const el = app.getByTestId('App');
+    expect(el).toBeInstanceOf(HTMLElement);
   });
 
-  it('app', ()=>{
-    let about = render(About());
-    let el = about.getByTestId("About");
-    expect(el.textContent).toContain('Lorem ipsum')
+  it('about', () => {
+    const about = render(About());
+    const el = about.getByTestId('About');
+    expect(el.textContent).toContain('Lorem ipsum');
   });
-
 });
